Integration of three_js and interactivevisualization was improved
* changed the load_csv() function.
According with the documentation:
index_col : int or sequence or False, default None
Column to use as the row labels of the DataFrame. If a sequence is given, a MultiIndex is used. If you have a malformed file with delimiters at the end of each line, you might consider index_col=False to force pandas to _not_ use the first column as the index (row names)

* currently we can choose all_jobs.csv file
* default index column is pandaid (teh ability to choose index should be implemented as GUI)
* the initial dataset are cleaned and normalized
* normalized procedure was changed - columns to normalized was added